requirements.txt
annoy==1.16.3
colorama==0.4.4
Cython==0.29.23
DateTime==4.3
fasteners==0.16.3
fire==0.4.0
gensim==4.1.2
h5py==2.10.0
joblib==1.1.0
lz4==3.1.10
monotonic==1.6
numpy==1.21.4
pandas==1.3.4
py-lru-cache==0.1.4
pymagnitude==0.1.143
pymagnitude-light==0.1.147
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
smart-open==5.2.1
termcolor==1.1.0
threadpoolctl==3.0.0
torch==0.4.1
tqdm==4.62.3
xxhash==2.0.2
zope.interface==5.4.0
